Redesign of the repository. Added pipeline training set maker.
