+++ .gitignore
@@ -135,4 +135,5 @@ dmypy.json
 # config files 
 env.sh
 pipelines.yml
+config.yml
 data-example/datasets.yml
