--- data-example/datasets.template.yml
@@ -1,5 +1,0 @@
-photo:
-  path: "<DATASETS_DIR>/photo"
-specz:
-  path: "<DATASETS_DIR>/specz"
-
