+++ data-example/mini_dataset/catalog_info.json
@@ -1,0 +1,8 @@
+{
+    "catalog_name": "mini_dataset",
+    "catalog_type": "object",
+    "total_rows": 40,
+    "epoch": "J2000",
+    "ra_column": "ra",
+    "dec_column": "dec"
+}
