+++ data-example/mini_dataset/partition_info.csv
@@ -1,0 +1,2 @@
+Norder,Npix,Dir
+0,4,0
