--- data-example/specz/catalog_info.json
@@ -1,8 +1,0 @@
-{
-    "catalog_name": "specz",
-    "catalog_type": "object",
-    "total_rows": 5,
-    "epoch": "J2000",
-    "ra_column": "ra",
-    "dec_column": "dec"
-}
