--- data-example/specz/partition_info.csv
@@ -1,2 +1,0 @@
-Norder,Dir,Npix,num_rows
-0,0,8,5
