+++ env.template.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+source `dirname $CONDA_EXE`/activate || { echo "Failed to activate Conda environment"; exit 1; }
+
 export APP_DIR="<APP_DIR>"
 export PIPELINES_DIR="<PIPELINES_DIR>"
 export DATASETS_DIR="<DATASETS_DIR>"
