--- hello_parsl/packages/parsl_config.py
@@ -1,10 +1,0 @@
-"""_summary_ """
-
-from parsl.config import Config
-
-
-def get_parsl_config(modname) -> Config:
-    mod = __import__(f"executors.{modname}", fromlist=["create_executor"])
-    _func = getattr(mod, "create_executor")
-    executor_instance = _func()
-    return Config(executors=[executor_instance], strategy=None)
