+++ pipelines.template.yml
@@ -1,15 +1,7 @@
-hello_parsl:
-  dname: 'Hello World with Parsl'
-  dir: '<PIPELINES_DIR>/hello_parsl'
+training_set_maker:
+  dname: 'Training Set Maker'
+  dir: '<PIPELINES_DIR>/training_set_maker'
   runner: 'bash'
   executable: 'run.sh'
   config: 'config.yml'
-  version: '0.1.0'
-
-cross_lsdb:
-  dname: 'LSDB Crossmatch'
-  dir: '<PIPELINES_DIR>/cross_lsdb'
-  runner: 'bash'
-  executable: 'run.sh'
-  config: 'config.yml'
-  version: '0.1.0'
+  version: '0.0.1'
