+++ training_set_maker/VERSION
@@ -1,0 +1,1 @@
+0.0.1
